restructured deps and packages
